Added enemy interaction sound effects [ROUGH DRAFT]
This is a rough draft of adding sound effects using SoundPlayer. I created the function AsyncPlay() in an attempt to use threads to make up for SoundPlayer's lack of asynchronous sound support, but to no avail. Playing sounds before other sounds may be as far as I can take this feature without adding external libraries/resources.

This version also contains several commented out lines of code that I am saving in this commit for future reference. The next commit will not include these.

Known bugs for this feature:
-  enemy_interact_1.wav plays for both the enemy and the boss when it should only play for the enemy.
- enemy_interact_2.wav does not play at all; it's one or the other.
- My attempt at a thread is either faulty or SoundPlayer's lack of support for asyncronous sound simply cannot be made up for by using threads.

Plans for next commit:
- Clean up unnecessary code
- Figure out if I need to remove AsyncPlay() or just reformat it
- Ensure that only enemy_interact_1.wav plays before the battle music before each enemy/boss encounter